chore: upgrade test folder structure
